agentcore-runtime/pyproject.toml
[project]
name = "agentcore-runtime-agent"
version = "1.0.0"
description = "AgentCore Runtime Agent with Strands Agents"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "strands-agents>=1.20.0",
    "boto3>=1.35.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
